### changed
- cleaned up bindgen
- default page size decreased to ~60Mb
- suppressed improper_ctypes

### added
- variable max_pages

+++ src/ringbuf.rs
@@ -5,7 +5,6 @@ use std::sync::{Arc, RwLock};
 use thiserror::Error;
 
 /// A thread safe ringbuf receiver
-///
 #[derive(Clone)]
 pub struct Reader {
     path: PathBuf,
@@ -17,18 +16,7 @@ pub struct Reader {
     _lock: Arc<fslock::LockFile>,
 }
 
-/** A thread safe ringbuf sender
-```rust
-    let (mut tx, mut rx) = new("test-spsc").unwrap();
-
-    let now = std::time::Instant::now();
-    let t = std::thread::spawn(move || {
-        for i in 0..50_000_000 {
-            tx.push(i.to_string()).unwrap();
-        }
-    });
-```
-**/
+/// A thread safe ringbuf sender
 #[derive(Clone)]
 pub struct Writer {
     path: PathBuf,
@@ -58,7 +46,6 @@ impl From<i32> for RingbufError {
     }
 }
 
-const TEMP_MAX_TOTAL_PAGES: usize = 4;
 const PAGE_EXT: &str = "page.bin";
 
 fn check_valid_page(entry: DirEntry) -> Option<usize> {
@@ -90,7 +77,7 @@ fn find_pages<P: AsRef<Path>>(path: P) -> Result<usize, RingbufError> {
     Ok(write_page_count)
 }
 
-pub fn new<P: Into<PathBuf>>(path: P) -> Result<(Writer, Reader), RingbufError> {
+pub fn new<P: Into<PathBuf>>(path: P, max_pages: usize) -> Result<(Writer, Reader), RingbufError> {
     let path = path.into();
     std::fs::create_dir_all(&path)?;
 
@@ -120,7 +107,7 @@ pub fn new<P: Into<PathBuf>>(path: P) -> Result<(Writer, Reader), RingbufError> 
             write_page_count: wp_count.clone(),
             write_page_no: 0,
             write_page: page.clone(),
-            max_total_pages: TEMP_MAX_TOTAL_PAGES,
+            max_total_pages: max_pages,
             _lock: _lock.clone(),
         },
         Reader {
@@ -129,7 +116,7 @@ pub fn new<P: Into<PathBuf>>(path: P) -> Result<(Writer, Reader), RingbufError> 
             read_page_no: 0,
             read_page: page,
             read_start_byte: 0,
-            max_total_pages: TEMP_MAX_TOTAL_PAGES,
+            max_total_pages: max_pages,
             _lock,
         },
     ))
@@ -156,6 +143,11 @@ impl Writer {
 
             *page_count += 1;
             self.write_page_no += 1;
+
+            // setting max_total_pages to zero implies an unbounded ringbuf / queue
+            if self.max_total_pages == 0 {
+                return Ok(());
+            }
 
             if *page_count >= self.max_total_pages {
                 std::fs::remove_file(
@@ -244,17 +236,20 @@ impl Reader {
 
 #[test]
 fn lock_test() {
-    let (_tx, _rx) = new("test-lock").unwrap();
-    assert!(new("test-lock").is_err());
+    let test_dir_path = "test-lock";
+    let (_tx, _rx) = new(test_dir_path, 2).unwrap();
+    assert!(new(test_dir_path, 2).is_err());
 
     drop(_tx);
     drop(_rx);
-    let (_tx, _rx) = new("test-lock").unwrap();
+    let (_tx, _rx) = new(test_dir_path, 2).unwrap();
+    std::fs::remove_dir_all(test_dir_path).unwrap();
 }
 
 #[test]
 fn seq_test() {
-    let (mut tx, mut rx) = new("test-seq").unwrap();
+    let test_dir_path = "test-seq";
+    let (mut tx, mut rx) = new(test_dir_path, 8).unwrap();
 
     let now = std::time::Instant::now();
     for i in 0..50_000_000 {
@@ -267,11 +262,14 @@ fn seq_test() {
     }
 
     eprintln!("took {} ms", now.elapsed().as_millis());
+
+    std::fs::remove_dir_all(test_dir_path).unwrap();
 }
 
 #[test]
 fn spsc_test() {
-    let (mut tx, mut rx) = new("test-spsc").unwrap();
+    let test_dir_path = "test-spsc";
+    let (mut tx, mut rx) = new(test_dir_path, 8).unwrap();
 
     let now = std::time::Instant::now();
     let t = std::thread::spawn(move || {
@@ -298,6 +296,6 @@ fn spsc_test() {
     let _ = t.join().unwrap();
 
     eprintln!("took {} ms", now.elapsed().as_millis());
+
+    std::fs::remove_dir_all(test_dir_path).unwrap();
 }
-// deleting pages on pop makes life much easier as opposed to deleting
-// old pages on push which might screw things up
